+++ starters/nextjs-starter-ts/pages/articles/index.tsx
@@ -1,17 +1,30 @@
 import {
   Article,
-  ArticleWithoutContent,
   PCCConvenienceFunctions,
   Site,
 } from "@pantheon-systems/pcc-react-sdk";
 import { NextSeo } from "next-seo";
-import { ArticleGrid } from "../../components/grid";
+import queryString from "query-string";
+import ArticleList from "../../components/article-list";
 import Layout from "../../components/layout";
-import PageHeader from "../../components/page-header";
-import Pagination from "../../components/pagination";
-import { usePagination } from "../../hooks/usePagination";
+import { PAGE_SIZE } from "../../constants";
 
-const PAGE_SIZE = 20;
+async function fetchNextPages(cursor?: string | null | undefined) {
+  const url = queryString.stringifyUrl({
+    url: "/api/utils/paginate",
+    query: {
+      pageSize: PAGE_SIZE,
+      cursor: cursor,
+    },
+  });
+
+  const response = await fetch(url);
+  const { data, cursor: newCursor } = await response.json();
+  return {
+    data,
+    newCursor,
+  };
+}
 
 interface Props {
   articles: Article[];
@@ -26,39 +39,18 @@ export default function ArticlesListTemplate({
   cursor,
   site,
 }: Props) {
-  const {
-    data: currentArticles,
-    onPageChange,
-    fetching,
-    currentPage,
-  } = usePagination({
-    cursor,
-    initialArticles: articles,
-    pageSize: PAGE_SIZE,
-  });
-
   return (
     <Layout>
       <NextSeo title="Articles" description="Articles" />
 
-      <section className="max-w-screen-3xl mx-auto px-4 pt-16 sm:w-4/5 md:w-3/4 lg:w-4/5 2xl:w-3/4">
-        <PageHeader title="Articles" />
-
-        <ArticleGrid
-          articles={currentArticles as ArticleWithoutContent[]}
-          site={site}
-        />
-
-        <div className="mt-4 flex flex-row items-center justify-center">
-          <Pagination
-            totalCount={totalCount}
-            pageSize={PAGE_SIZE}
-            currentPage={currentPage}
-            onChange={onPageChange}
-            disabled={fetching}
-          />
-        </div>
-      </section>
+      <ArticleList
+        headerText="Articles"
+        articles={articles}
+        cursor={cursor}
+        totalCount={totalCount}
+        fetcher={fetchNextPages}
+        site={site}
+      />
     </Layout>
   );
 }