+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pantheon-systems/pcc-cli
 
+## 3.12.1
+
+### Patch Changes
+
+- @pantheon-systems/pcc-sdk-core@3.12.1
+
+## 3.12.0
+
+### Patch Changes
+
+- @pantheon-systems/pcc-sdk-core@3.12.0
+
 ## 3.11.3
 
 ### Patch Changes