+++ packages/react-sdk/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @pantheon-systems/pcc-react-sdk
 
+## 3.12.1
+
+### Patch Changes
+
+- b587dda: New CSS classes: pantheon-img-container-breakboth and
+  pantheon-img-container-alone
+  - @pantheon-systems/pcc-sdk-core@3.12.1
+
+## 3.12.0
+
+### Minor Changes
+
+- 33368d3: Image captions are now rendered by default, but can be opted out by
+  using \_\_experimentalFlags.renderImageCpations in ArticleRenderer. Image
+  containers and captions are now targetable with CSS classes.
+
+### Patch Changes
+
+- @pantheon-systems/pcc-sdk-core@3.12.0
+
 ## 3.11.3
 
 ### Patch Changes